requirements.txt
# VLA Training Requirements
# =========================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & HuggingFace
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0
bitsandbytes>=0.42.0
safetensors>=0.4.0
tokenizers>=0.15.0

# DeepSpeed (for distributed training)
deepspeed>=0.13.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Reinforcement Learning (optional)
gymnasium>=0.29.0
# gymnasium[mujoco]>=0.29.0  # Uncomment for MuJoCo envs

# Visualization & Logging
matplotlib>=3.7.0
tqdm>=4.65.0
wandb>=0.16.0
tensorboard>=2.15.0

# LeRobot (optional, for robot datasets)
# lerobot>=0.1.0  # Install separately: pip install lerobot

# Flash Attention (optional, for faster attention)
# flash-attn>=2.5.0  # Install separately with CUDA

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Additional utilities
einops>=0.7.0
omegaconf>=2.3.0
hydra-core>=1.3.0
